requirements.txt
asgiref
certifi
chardet
coreapi
coreschema
Django
django-cors-headers
django-rest-passwordreset
djangorestframework
djangorestframework-simplejwt
drf-yasg
idna
inflection
itypes
Jinja2
MarkupSafe
packaging
PyJWT
pyparsing
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
uritemplate
urllib3
